#8618: move Handles, Cstring and TupleToArray into org.eolang.sys
They serve posix and win32 alone, yet standing in org.eolang they read
as part of the runtime's public surface, next to Bytes and PhDefault.
Syscall stays behind under a puzzle of its own: sixty-two adapters
import it, and moving it here as well would push the diff past the
two-hundred-line limit.
