+++ eo-runtime/src/main/java/org/eolang/posix/InetAddrSyscall.java
@@ -6,11 +6,11 @@ package org.eolang.posix;
 
 import com.sun.jna.Native;
 import java.util.Collections;
-import org.eolang.Cstring;
 import org.eolang.Data;
 import org.eolang.PhDefault;
 import org.eolang.Phi;
 import org.eolang.Syscall;
+import org.eolang.sys.Cstring;
 
 /**
  * The 'inet_addr' syscall.
