+++ eo-runtime/src/main/java/org/eolang/posix/MkdirSyscall.java
@@ -4,11 +4,11 @@
  */
 package org.eolang.posix;
 
-import org.eolang.Cstring;
 import org.eolang.Data;
 import org.eolang.Int;
 import org.eolang.Phi;
 import org.eolang.Syscall;
+import org.eolang.sys.Cstring;
 
 /**
  * Mkdir syscall.
