+++ eo-runtime/src/main/java/org/eolang/posix/OpendirSyscall.java
@@ -5,11 +5,11 @@
 package org.eolang.posix;
 
 import com.sun.jna.Pointer;
-import org.eolang.Cstring;
 import org.eolang.Data;
-import org.eolang.Handles;
 import org.eolang.Phi;
 import org.eolang.Syscall;
+import org.eolang.sys.Cstring;
+import org.eolang.sys.Handles;
 
 /**
  * Opendir syscall.
