+++ eo-runtime/src/main/java/org/eolang/posix/ReaddirSyscall.java
@@ -7,11 +7,11 @@ package org.eolang.posix;
 import com.sun.jna.Platform;
 import com.sun.jna.Pointer;
 import org.eolang.Data;
-import org.eolang.Handles;
 import org.eolang.Int;
 import org.eolang.PhDefault;
 import org.eolang.Phi;
 import org.eolang.Syscall;
+import org.eolang.sys.Handles;
 
 /**
  * Readdir syscall.
