+++ eo-runtime/src/main/java/org/eolang/posix/SymlinkSyscall.java
@@ -4,10 +4,10 @@
  */
 package org.eolang.posix;
 
-import org.eolang.Cstring;
 import org.eolang.Data;
 import org.eolang.Phi;
 import org.eolang.Syscall;
+import org.eolang.sys.Cstring;
 
 /**
  * Symlink syscall.
