+++ eo-runtime/src/main/java/org/eolang/sys/TupleToArray.java
@@ -3,9 +3,12 @@
  * SPDX-License-Identifier: MIT
  */
 
-package org.eolang;
+package org.eolang.sys;
 
 import java.util.function.Supplier;
+import org.eolang.Dataized;
+import org.eolang.ExFailure;
+import org.eolang.Phi;
 
 /**
  * Convert {@code EOtuple} of arguments to Java array.
