+++ eo-runtime/src/main/java/org/eolang/win32/FindCloseFuncCall.java
@@ -6,11 +6,11 @@ package org.eolang.win32;
 
 import com.sun.jna.Pointer;
 import org.eolang.Data;
-import org.eolang.Handles;
 import org.eolang.Int;
 import org.eolang.PhDefault;
 import org.eolang.Phi;
 import org.eolang.Syscall;
+import org.eolang.sys.Handles;
 
 /**
  * The kernel32 FindClose function call.
