+++ eo-runtime/src/main/java/org/eolang/win32/GetFileAttributesFuncCall.java
@@ -5,11 +5,11 @@
 package org.eolang.win32;
 
 import com.sun.jna.WString;
-import org.eolang.Cstring;
 import org.eolang.Data;
 import org.eolang.PhDefault;
 import org.eolang.Phi;
 import org.eolang.Syscall;
+import org.eolang.sys.Cstring;
 
 /**
  * The kernel32 GetFileAttributesW function call.
