+++ eo-runtime/src/main/java/org/eolang/win32/UnlinkFuncCall.java
@@ -5,10 +5,10 @@
 package org.eolang.win32;
 
 import com.sun.jna.WString;
-import org.eolang.Cstring;
 import org.eolang.Data;
 import org.eolang.Phi;
 import org.eolang.Syscall;
+import org.eolang.sys.Cstring;
 
 /**
  * The msvcrt _unlink function call.
