+++ eo-runtime/src/test/java/org/eolang/sys/CstringTest.java
@@ -2,8 +2,12 @@
  * SPDX-FileCopyrightText: Copyright (c) 2016-2026 Objectionary.com
  * SPDX-License-Identifier: MIT
  */
-package org.eolang;
+package org.eolang.sys;
 
+import org.eolang.Data;
+import org.eolang.ExFailure;
+import org.eolang.Expect;
+import org.eolang.PhTerminator;
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
 import org.junit.jupiter.api.Assertions;
